Add cli flag to enable events on workers
Add `--enable-events` CLI argument, which causes
celery-prometheus-exporter to periodically enable events on workers.

This is useful because it allows running celery workers with events
disabled, until `celery-prometheus-exporter` is deployed, at which time
events get enabled on the workers.

Celery Flower does the same thing. Except there, this is the default
behavior. In this case I made it opt-in to keep previous behavior.