Added Histogram 'runtime' metric for successfully-completed tasks.
runtime is the time it took to execute the task using the pool.
(Starting from the task is sent to the worker pool, and ending
when the pool result handler callback is called).

This metric is not recorded by state, as Celery only provides
the runtime metric for successfully completed tasks.